wip: make end to end scripts more flexible
 - being improving demo documentation in regards to the scripts
 - add help text to scripts
 - add more flags to teardown and start script to allow manipulation
   of just the network

+++ demo/scripts/teardown.sh
@@ -14,15 +14,59 @@ DEMO_DOCKER_COMPOSE=${DEMO_SCRIPT_DIR}/../docker-compose.yml
 SCRIPT_DIR=${DEMO_SCRIPT_DIR}/../../utils/docker-compose/scripts
 . ${SCRIPT_DIR}/lib/common.sh
 
-COMPOSE_IGNORE_ORPHANS=true docker-compose -f ${DEMO_DOCKER_COMPOSE} kill && COMPOSE_IGNORE_ORPHANS=true docker-compose -f ${DEMO_DOCKER_COMPOSE} down
+help(){
+   echo "$(basename $0) [options]
+
+   This script, by default, will teardown all of the auction demo components, but will not prune
+   volumes or delete the chaincode images that have been created for the mockcc and the auctioncc
+
+   options:
+       --clean-slate:
+           As part of the teardown, prune all unused docker volumes and delete mockcc and auctioncc
+           images.
+       --help,-h:
+           Print this help screen.
+       --network-only,-n:
+           The script will not teardown the fabric-gateway or the UI. It will only teardown the
+           FPC Network as definied in \$FPC_PATH/utils/docker-compose.
+    "
+}
+
+CLEAN_SLATE=false
+NETWORK_ONLY=false
+for var in "$@"; do
+    case "$var" in
+        "--clean-slate")
+            BUILD_CHAINCODE=true
+            ;;
+        "-n"|"--network-only")
+            NETWORK_ONLY=true
+            ;;
+        "-h"|"--help"|*)
+            help
+            exit
+            ;;
+    esac
+    shift
+done
+
+if [ ! $NETWORK_ONLY ]; then
+    COMPOSE_IGNORE_ORPHANS=true docker-compose -f ${DEMO_DOCKER_COMPOSE} kill && COMPOSE_IGNORE_ORPHANS=true docker-compose -f ${DEMO_DOCKER_COMPOSE} down
+fi
 
 
-if [ "$1" == '--clean-slate' ]; then
+if [ $CLEAN_SLATE ]; then
     "${SCRIPT_DIR}/teardown.sh" --clean-slate
     images=$(docker images dev_*mockcc-* -q)
     if [ ! -z "${images}" ]; then
         docker rmi "${images}"
     fi
-else
-    "${SCRIPT_DIR}/teardown.sh"
+
+    images=$(docker images dev_*auctioncc-* -q)
+    if [ ! -z "${images}" ]; then
+        docker rmi "${images}"
+    fi
+    exit
 fi
+
+"${SCRIPT_DIR}/teardown.sh"
